% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPairs.R
\name{getPairs}
\alias{getPairs}
\title{Gets Pairs from Distances}
\usage{
getPairs(dist, k)
}
\arguments{
\item{dist}{The distance matrix describing the distance between locations.}

\item{k}{The number of closest neighbors to be located for each lcoation.}
}
\value{
A named list.  Each element of the list corresponds to a particular
location, and the value at element i is the k closest locations to location
i.
}
\description{
For each location, we wish to determine the k closest locations.  This
function takes the distance matrix and computes the returns a list of the
k closest locations to each individual location.
}

