% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getUniquePairs.R
\name{getUniquePairs}
\alias{getUniquePairs}
\title{Get Unique Pairs}
\usage{
getUniquePairs(pairs)
}
\arguments{
\item{pairs}{The pairs list object, as returned by ?getPairs.}
}
\value{
data.frame with columns loc1 and loc2.  This data.frame will have
no duplicates, and describes all the pairs that need to be computed.
}
\description{
For the pairwise SNHT, many difference time-series must be computed.
However, if location i is used for location j, then it's very likely that
location j will be used for location i.  Thus, the pairs object will likely
have many duplicate pairs.  To save computation time, this function finds
which pairs are unique.
}

