% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSNHT.R
\name{plotSNHT}
\alias{plotSNHT}
\title{Plot SNHT}
\usage{
plotSNHT(data, stat, time = NULL, alpha = NULL)
}
\arguments{
\item{data}{The vector of time series observations that were input to the
snht function.}

\item{stat}{A data.frame as returned by the snht function.}

\item{time}{If the observations in data are not equally spaced, then this
vector will specify the times of the observations.  This object should be
numeric or should be able to be coerced to numeric.}

\item{alpha}{The confidence level for the SNHT test.  Note, though, that
multiple tests are being performed and that is NOT accounted for in this
function.}
}
\value{
No object is returned, but a plot is instead generated.
}
\description{
Function to plot the result of the SNHT fit
}

