% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snht-package.R
\docType{package}
\name{snht-package}
\alias{snht-package}
\title{Robust and Non-Robust Standard Normal Homogeneity Test}
\description{
Computes test statistics for the SNHT and robust SNHT test. Additionally,
users may supply a custom function for estimating the mean and standard
deviation, and this function will be used for computing the test statistic.
}
\details{
The only function you're likely to need from \pkg{snht} is
\code{\link{snht}}. Otherwise refer to the vignettes.
}

